+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.4 - 2020-09-20
+### Added
+- Support for non-standard directions.txt file.
+- Added getFareAttributes to README
+
 ## [2.0.3 - 2020-09-14
 ### Fixed
 - Fix for querying for null